Properly prepend the expando with "batman"
This was discovered in IE8 testing where the expando became a string of
digits longer than 10 characters. It turns out that IE8 doesn't like to
look up properties whose names are a string of digits longer than 10
characters on SELECT nodes.  ie. node["12345678901"] blows up, but
node["a12345678901"] is fine, where node is a SELECT element.