Only add the keys actually in the collection.
Otherwise, this traverses the Object prototype as well, which wreaks
havoc with with `Batman.SimpleHash::set` (e.g. the key `constructor`
actually retrieves something from `@_storage`, but that thing is not
array-like, so the implicit `pair[0] for pair in Object::constructor`
dies).

It probably bears looking at the other instances of `for ... of` and
ensuring that they have the intended behavior.