Don't assume that an associated model has been loaded
This necessitated the remove of Association.relatedModel since we don't
know what that'll be, and involved making Batman.Association.Collection
a singleton which tracks all associations (rather than each model
tracking its own associations).

The model code now looks up associations info in the collection
singleton when needed. Models are looked up in Batman.currentApp.

The tests had to be updated to operate under an app so that
Batman.currentApp would be defined.