Don't apply an undefined value during a belongsTo save
This is a problem when an association proxy doesn't have a defined value for the key. It can cause a property that was set elsewhere to become unset because the proxy doesn't have the property.

ie. Saving a new article which has gotten a blog_id from some binding in the view, but its blog proxy doesn't yet have an id. That blog_id property shouldn't be set to undefined.